Add support for create and delete event operations.
Two changes:

1. If a timestamp of class Time is passed to an event set add/remove method,
it will create either a "create" (add) or "delete" (remove) event for
the instance in that event set. Per the API docs, this won't work if
the instance doesn't exist at the time specified.

2. Fauna::Event objects can be saved and deleted, causing the event set
they belong to to be updated accordingly. Deleting an event is pretty
straightforward (find the event in an instance or set's events and delete
it, and it vanishes).

There's no use case that I can see for saving an event that you've
retrieved from an event set (the timestamp of Fauna::Event isn't writable),
but you could instantiate a new Fauna::Event with the right parameters
(including a timestamp attribute in usecs, not Time), and save that to
create a new event in an instance's event history.